migrate useful env vars to config(.cpu).yaml
This helps with the docker containers since docker does not allow
changing of env vars after creation of a container. Having this config
file (also in persistent storage) helps with changing of configuration
without the hassle of re-creating it, which with the AppAPI
architecture, results in a container with a similar state as before.

Signed-off-by: Anupam Kumar <kyteinsky@gmail.com>