Fix: Sources that are decoded as only spaces fail to index (#157)
I noticed that some of the documents that fail to get indexed were
because they get decoded as a lot of empty characters which the
embedding model fails to embed. This just changes the check for an empty
document to strip whitespace characters during the check. I wrote a
quick fix that worked on my instance and caused the error to disappear.

---------

Signed-off-by: Lukas <lukas@lschaefer.xyz>